+++ NEWS.md
@@ -1,3 +1,4 @@
+
 # version 0.7.0
 
 - moving block bootstrap in `ridge2f`, `basicf` and `loessf`, in addition to circular block bootstrap from 0.6.2